+++ app/[domain]/(default)/timelines/content.tsx
@@ -1,13 +1,19 @@
 'use client';
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
+import { useAppKitAccount } from '@reown/appkit/react';
+import { match } from 'ts-pattern';
 
-import { ASSET_PREFIX, LEMONADE_FEED_ADDRESS } from '$lib/utils/constants';
-import { Segment } from '$lib/components/core';
+import { LEMONADE_FEED_ADDRESS } from '$lib/utils/constants';
+import { Button, modal, Segment } from '$lib/components/core';
 import { FeedPosts } from '$lib/components/features/lens-feed/FeedPosts';
-import { PostComposer } from '$lib/components/features/lens-feed/PostComposer';
+import { PostComposer, PostLocked } from '$lib/components/features/lens-feed/PostComposer';
 import { useAccount, usePost } from '$lib/hooks/useLens';
 import { Timeline } from '$lib/components/features/lens-feed/Timeline';
+import { useQuery } from '$lib/graphql/request';
+import { GetSpaceDocument, Space } from '$lib/graphql/generated/backend/graphql';
+import { useLemonhead } from '$lib/hooks/useLemonhead';
+import { ConnectWallet } from '$lib/components/features/modals/ConnectWallet';
 
 type Tab = 'Following' | 'Lemonade' | 'LemonHeads' | 'Global';
 
@@ -15,6 +21,11 @@ export function HomePageContent() {
   const router = useRouter();
   const { createPost } = usePost();
   const { account } = useAccount();
+  const { isConnected } = useAppKitAccount();
+  const { data } = useLemonhead();
+
+  const { data: dataGetSpace } = useQuery(GetSpaceDocument, { variables: { slug: 'lemonheads' } });
+  const feedAddress = (dataGetSpace?.getSpace as Space)?.lens_feed_id || LEMONADE_FEED_ADDRESS;
 
   const [tab, setTab] = useState<Tab>('Lemonade');
 
@@ -43,37 +54,52 @@ export function HomePageContent() {
         className="bg-transparent w-full md:w-fit overflow-auto no-scrollbar"
       />
 
-      {tab !== 'LemonHeads' && <PostComposer onPost={onPost} showFeedOptions />}
+      {match(tab)
+        .with('LemonHeads', () => (
+          <PostComposer
+            onPost={onPost}
+            showFeedOptions={false}
+            renderLock={match(isConnected)
+              .with(false, () => (
+                <PostLocked title="Posting is Locked" subtitle="Connect wallet to start posting on Lemonade.">
+                  <Button
+                    variant="secondary"
+                    size="sm"
+                    className="rounded-full"
+                    onClick={() => modal.open(ConnectWallet, { props: { onConnect: () => {} } })}
+                  >
+                    Connect Wallet
+                  </Button>
+                </PostLocked>
+              ))
+              .otherwise(() => {
+                if (!data || (data && data.tokenId == 0)) {
+                  return (
+                    <PostLocked title="Posting is Locked" subtitle="Claim Lemonhead to start posting">
+                      <Button
+                        variant="secondary"
+                        className="rounded-full"
+                        onClick={() => router.push('/lemonheads/mint')}
+                        size="sm"
+                      >
+                        Claim Lemonhead
+                      </Button>
+                    </PostLocked>
+                  );
+                }
+              })}
+          />
+        ))
+        .otherwise(() => (
+          <PostComposer onPost={onPost} showFeedOptions />
+        ))}
 
-      {tab === 'Following' && <Timeline account={account?.address} onSelectPost={onSelectPost} />}
-
-      {tab === 'Lemonade' && <FeedPosts feedAddress={LEMONADE_FEED_ADDRESS} onSelectPost={onSelectPost} />}
-
-      {tab === 'LemonHeads' && (
-        <div className="flex gap-6 items-center flex-col mt-[108px]">
-          <div
-            className="size-[184px] p-2 rounded-lg"
-            style={{
-              background: `url(${ASSET_PREFIX}/assets/images/lemonheads-bg.png)`,
-            }}
-          >
-            <img src={`${ASSET_PREFIX}/assets/images/lemonheads.gif`} alt="LemonHeads" className="rounded-md" />
-          </div>
-
-          <div className="space-y-2">
-            <h1 className="text-xl font-semibold text-center">LemonHeads Social Feed</h1>
-            <p className="text-accent-400 text-center">Coming Soon!</p>
-            <div>
-              <p className="text-secondary text-center">
-                An exclusive space for LemonHeads to post, connect & socialize.
-              </p>
-              <p className="text-secondary text-center">Only LemonHead owners get access.</p>
-            </div>
-          </div>
-        </div>
-      )}
-
-      {tab === 'Global' && <FeedPosts global={true} onSelectPost={onSelectPost} />}
+      {match(tab)
+        .with('Following', () => <Timeline account={account?.address} onSelectPost={onSelectPost} />)
+        .with('Lemonade', () => <FeedPosts feedAddress={LEMONADE_FEED_ADDRESS} onSelectPost={onSelectPost} />)
+        .with('LemonHeads', () => <FeedPosts feedAddress={feedAddress} onSelectPost={onSelectPost} />)
+        .with('Global', () => <FeedPosts global={true} onSelectPost={onSelectPost} />)
+        .exhaustive()}
     </div>
   );
 }