+++ lib/components/features/lemonheads/LemonheadsGallery.tsx
@@ -7,36 +7,57 @@ import { GetMintedsDocument, Minted_OrderBy, OrderDirection } from '$lib/graphql
 import { Skeleton, Button } from '$lib/components/core';
 import { TokenCard } from './TokenCard';
 
+const LIMIT = 20;
+
 export function LemonheadsGallery() {
   const scrollContainerRef = React.useRef<HTMLDivElement>(null);
+
+  const [canFetchMore, setCanFetchMore] = React.useState(true);
+
   const [canScrollLeft, setCanScrollLeft] = React.useState(false);
   const [canScrollRight, setCanScrollRight] = React.useState(false);
+  const [mounted, setMounted] = React.useState(false);
 
-  const { data, loading } = useQuery(GetMintedsDocument, {
-    variables: {
-      first: 20,
-      orderBy: Minted_OrderBy.TokenId,
-      orderDirection: OrderDirection.Desc
-    }
-  }, lemonheadsClient);
+  const { data, loading, fetchMore } = useQuery(
+    GetMintedsDocument,
+    {
+      variables: {
+        first: LIMIT,
+        skip: 0,
+        orderBy: Minted_OrderBy.TokenId,
+        orderDirection: OrderDirection.Desc,
+      },
+    },
+    lemonheadsClient,
+  );
+
+  const onLoadMore = () => {
+    fetchMore({
+      variables: { first: LIMIT, skip: data?.minteds?.length },
+      updateQuery: (existing, incomes) => {
+        if (existing?.minteds && incomes?.minteds?.length) {
+          const minteds = existing?.minteds || [];
+          return { __typename: 'Query', minteds: [...minteds, ...incomes.minteds] };
+        }
+
+        if (!incomes.minteds.length) setCanFetchMore(false);
+
+        return existing;
+      },
+    });
+  };
 
   const checkScrollability = () => {
     if (scrollContainerRef.current) {
       const { scrollLeft, scrollWidth, clientWidth } = scrollContainerRef.current;
       setCanScrollLeft(scrollLeft > 0);
       setCanScrollRight(scrollLeft < scrollWidth - clientWidth);
-    }
-  };
 
-  const scrollLeft = () => {
-    if (scrollContainerRef.current) {
-      scrollContainerRef.current.scrollBy({ left: -240, behavior: 'smooth' });
-    }
-  };
+      const isNearEnd = scrollLeft + clientWidth + 240 >= scrollWidth;
 
-  const scrollRight = () => {
-    if (scrollContainerRef.current) {
-      scrollContainerRef.current.scrollBy({ left: 240, behavior: 'smooth' });
+      if (isNearEnd && !loading && mounted && data?.minteds.length && canFetchMore) {
+        onLoadMore();
+      }
     }
   };
 
@@ -44,24 +65,30 @@ export function LemonheadsGallery() {
     const container = scrollContainerRef.current;
     if (container) {
       checkScrollability();
+      setMounted(true);
       container.addEventListener('scroll', checkScrollability);
       return () => container.removeEventListener('scroll', checkScrollability);
     }
-  }, [data]);
+  }, [scrollContainerRef.current, data?.minteds.length, canFetchMore, loading]);
 
-  if (loading) {
-    return (
-      <div className="flex overflow-x-auto no-scrollbar gap-0 rounded-md overflow-hidden border-2 border-card-border divide-card-border divide-x-2">
-        {Array.from({ length: 20 }).map((_, index) => (
-          <Skeleton key={index} className="min-w-60 h-60 rounded-none" />
-        ))}
-      </div>
-    );
-  }
+  const scrollLeft = () => {
+    if (scrollContainerRef.current) {
+      scrollContainerRef.current.scrollBy({ left: -240, behavior: 'smooth' });
+    }
+  };
 
-  const minteds = data?.minteds || [];
+  const scrollRight = async () => {
+    if (scrollContainerRef.current) {
+      const { scrollLeft, scrollWidth, clientWidth } = scrollContainerRef.current;
+      const isNearEnd = scrollLeft + clientWidth + 240 >= scrollWidth;
 
-  if (!minteds.length) return null;
+      if (isNearEnd && !loading && mounted && data?.minteds.length && canFetchMore) {
+        onLoadMore();
+      }
+
+      scrollContainerRef.current.scrollBy({ left: 240, behavior: 'smooth' });
+    }
+  };
 
   return (
     <div className="relative">
@@ -69,12 +96,14 @@ export function LemonheadsGallery() {
         ref={scrollContainerRef}
         className="flex overflow-x-auto no-scrollbar gap-0 rounded-md border-2 border-card-border divide-card-border divide-x-2"
       >
-        {minteds.map((minted) => (
-          <TokenCard
-            key={minted.tokenId}
-            tokenId={minted.tokenId.toString()}
-          />
-        ))}
+        {data?.minteds?.map((minted, idx) => <TokenCard key={idx} tokenId={minted.tokenId.toString()} />)}
+        {loading && (
+          <div className="flex w-full h-60 divide-x-2">
+            {Array.from({ length: 5 }).map((_, index) => (
+              <Skeleton key={`loading-${index}`} className="size-60 aspect-square rounded-none" animate />
+            ))}
+          </div>
+        )}
       </div>
       {canScrollLeft && (
         <Button